element components are moved inside main component
